requirements.txt
flask
gunicorn
python-dotenv
google-generativeai
flask-cors
